change message name from COMPANION_STATUS to COMPANION_PROCESS_STATUS
